+++ src/type/TodoData.ts
@@ -1,0 +1,5 @@
+export type TodoData = {
+    id: number;
+    title: string;
+    completed: boolean;
+};
